backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
iniconfig==1.1.1
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.11.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
